+++ tests/unit/driver/core/test_cisco_nxos.py
@@ -2,13 +2,13 @@ import re
 
 import pytest
 
-from scrapli.driver.core.cisco_iosxe.driver import PRIVS
+from scrapli.driver.core.cisco_nxos.driver import PRIVS
 
 
 @pytest.mark.parametrize(
     "priv_pattern",
-    [("configuration", "csr1000v(config)#"), ("configuration", "csr1000v(conf-ssh-pubkey-data)#")],
-    ids=["base_prompt", "ssh_key_prompt"],
+    [("privilege_exec", "switch# "), ("configuration", "switch(config)# ")],
+    ids=["privilege_exec", "configuration"],
 )
 def test_prompt_patterns(priv_pattern):
     priv_level_name = priv_pattern[0]
@@ -18,15 +18,15 @@ def test_prompt_patterns(priv_pattern):
     assert match
 
 
-def test_on_open_on_close(mocked_iosxe_driver):
+def test_on_open_on_close(mocked_nxos_driver):
     channel_input_1 = "\n"
-    channel_output_1 = "\n3560CX#"
+    channel_output_1 = "\nswitch#"
     channel_input_2 = "terminal length 0"
-    channel_output_2 = "\n3560CX#"
+    channel_output_2 = "\nswitch#"
     channel_input_3 = "terminal width 512"
-    channel_output_3 = "\n3560CX#"
+    channel_output_3 = "\nswitch#"
     channel_input_4 = "\n"
-    channel_output_4 = "\n3560CX#"
+    channel_output_4 = "\nswitch#"
     channel_input_5 = "exit"
     channel_output_5 = ""
     test_operations = [
@@ -36,8 +36,8 @@ def test_on_open_on_close(mocked_iosxe_driver):
         (channel_input_4, channel_output_4),
         (channel_input_5, channel_output_5),
     ]
-    # mocked iosxe driver already calls `.open()` so we are just testing that the open commands
+    # mocked nxos driver already calls `.open()` so we are just testing that the open commands
     # for both of these methods get sent/read back from the channel... this is mostly to ensure
     # that any change to the open/close methods are noticed and also for vanity coverage :)
-    conn = mocked_iosxe_driver(test_operations)
+    conn = mocked_nxos_driver(test_operations)
     conn.close()