+++ tests/unit/driver/test_generic_driver.py
@@ -31,8 +31,14 @@ TEST_DATA_PATH = f"{Path(scrapli.__file__).parents[1]}/tests/unit/test_data"
             TypeError,
             "`send_commands` expects a list of strings, got <class 'str'>. to send a single command use the `send_command` method instead.",
         ),
+        (
+            "send_commands_from_file",
+            False,
+            TypeError,
+            "`send_commands_from_file` expects a string path to file, got <class 'bool'>",
+        ),
     ],
-    ids=["send_command", "send_commands",],
+    ids=["send_command", "send_commands", "send_commands_from_file"],
 )
 def test_send_commands_exceptions(attr_setup, mocked_generic_driver):
     method = attr_setup[0]
@@ -115,6 +121,41 @@ def test_send_commands(mocked_generic_driver):
     assert outputs[1].result == channel_output_2
 
 
+def test_send_commands_stop_on_failed(mocked_generic_driver):
+    channel_input_1 = "show ip access-lists"
+    channel_output_1 = "% invalid command\n3560CX#"
+    channel_input_2 = "show ip access-lists"
+    channel_output_2 = """Extended IP access list ext_acl_fw
+            10 deny ip 0.0.0.0 0.255.255.255 any
+            20 deny ip 10.0.0.0 0.255.255.255 any
+            30 deny ip 100.64.0.0 0.63.255.255 any (2 matches)
+            40 deny ip 127.0.0.0 0.255.255.255 any
+            50 deny ip 169.254.0.0 0.0.255.255 any
+            60 deny ip 172.16.0.0 0.15.255.255 any
+            70 deny ip 192.0.0.0 0.0.0.255 any
+            80 deny ip 192.0.2.0 0.0.0.255 any
+            90 deny ip 192.168.0.0 0.0.255.255 any
+            100 deny ip 198.18.0.0 0.1.255.255 any
+            110 deny ip 198.51.100.0 0.0.0.255 any
+            120 deny ip 203.0.113.0 0.0.0.255 any
+            130 deny ip 224.0.0.0 15.255.255.255 any
+            140 deny ip 240.0.0.0 15.255.255.255 any
+3560CX#"""
+    test_operations = [
+        (channel_input_1, channel_output_1),
+        (channel_input_2, channel_output_2),
+    ]
+    conn = mocked_generic_driver(test_operations)
+    outputs = conn.send_commands(
+        [channel_input_1, channel_input_2],
+        strip_prompt=False,
+        failed_when_contains=["% invalid command"],
+        stop_on_failed=True,
+    )
+    assert outputs[0].result == channel_output_1
+    assert len(outputs) == 1
+
+
 def test_send_commands_from_file(mocked_generic_driver):
     channel_input_1 = "show ip access-lists"
     channel_output_1 = """Extended IP access list ext_acl_fw
@@ -175,3 +216,12 @@ def test_send_inputs_interact(mocked_generic_driver):
     conn = mocked_generic_driver(test_operations)
     output = conn.send_interactive(interact)
     assert output.result == "clear logg\nClear logging buffer [confirm]\n3560CX#"
+
+
+def test_get_prompt(mocked_generic_driver):
+    channel_input_1 = "\n"
+    channel_output_1 = "\n3560CX#"
+    test_operations = [(channel_input_1, channel_output_1)]
+    conn = mocked_generic_driver(test_operations)
+    output = conn.get_prompt()
+    assert output == "3560CX#"