+++ tests/unit/transport/test_telnet.py
@@ -2,7 +2,7 @@ import sys
 
 import pytest
 
-from scrapli.exceptions import ScrapliAuthenticationFailed
+from scrapli.exceptions import ConnectionNotOpened, ScrapliAuthenticationFailed
 from scrapli.transport import TelnetTransport
 from scrapli.transport.telnet import ScrapliTelnet
 
@@ -36,3 +36,16 @@ def test_keepalive_standard():
     with pytest.raises(NotImplementedError) as exc:
         conn._keepalive_standard()
     assert str(exc.value) == "No 'standard' keepalive mechanism for telnet."
+
+
+@pytest.mark.parametrize(
+    "method_name", ["read", "write", "set_timeout"], ids=["read", "write", "set_timeout"],
+)
+def test_requires_open(method_name):
+    conn = TelnetTransport("localhost")
+    method = getattr(conn, method_name)
+    with pytest.raises(ConnectionNotOpened):
+        if method_name == "write":
+            method("blah")
+        else:
+            method()